requirements-windows.txt
# requirements-windows.txt — WSL/Windows Training Environment
# Install: pip install -r requirements-windows.txt
#
# Prerequisites:
#   - Python 3.12
#   - CUDA toolkit 12.x (via apt, NOT the nvidia-driver)
#   - ffmpeg (via apt)
#   - yt-dlp (via apt or pip)
#
# PyTorch must be installed separately with CUDA:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Core ML / Training
transformers>=4.36
accelerate
peft                          # LoRA / QLoRA adapters
trl                           # SFTTrainer for QLoRA fine-tuning
datasets                      # HuggingFace datasets
bitsandbytes                  # 4-bit quantization (QLoRA) + 8-bit optimizers
sentencepiece
protobuf

# Audio Preprocessing
soundfile
noisereduce
pyloudnorm
pydub
demucs                        # Source separation (CUDA-accelerated)
silero-vad
pyannote-audio                # Speaker diarization (requires HF token)
inaSpeechSegmenter            # Speech/music classification

# Evaluation
jiwer                         # WER/CER computation
sacrebleu                     # SacreBLEU + chrF++
unbabel-comet                 # COMET neural translation quality

# Utilities
numpy
pandas
tqdm
